==> _workspace/01-get-api-data-noodletracker.ipynb <==
### Get TikTok data

import pandas as pd
import datetime as dt
import numpy as np
import requests
from datetime import timedelta

today = (dt.datetime.today() + timedelta(days=30)).strftime('%Y-%m-%d')
today

payload={}
headers = {
  'X-API-KEY': 'QT8K1SQ333HurjoXkTVKOchKswsI83cS'
}

count = '30'
secUid = 'MS4wLjABAAAAn982fxeWVi0MRU_Q0Lg0mVJ9AmAZ-GvWgjy9rSGReFPs_xkQSFkKoUnkuWTVhryi'

rng = pd.date_range(start='2021-03-01', end = today, freq='M')
df = pd.DataFrame({ 'date' : rng }, )

df['cursor'] = (pd.to_datetime(df['date']) - dt.datetime(1970,1,1)).dt.total_seconds().round().astype(str).str.replace('.0', '', regex=False) + '000'

cursors = df['cursor'].to_list()

datas = []

for c in cursors:
    url = f"https://api.tikapi.io/public/posts?count={count}&secUid={secUid}&cursor={c}"
    r = requests.request("GET", url, headers=headers, data=payload)
    datas.append(r.json())

len(datas)

def simple_dict(tiktok_dict):
    to_return = {}
    to_return['user_name'] = d['author']['uniqueId']
    to_return['video_id'] = d['id']
    to_return['video_desc'] = d['desc']
    to_return['video_time'] = d['createTime']
    to_return['video_duration'] = d['video']['duration']
    to_return['n_shares'] = d['stats']['shareCount']
    to_return['n_likes'] = d['stats']['diggCount']
    to_return['n_comments'] = d['stats']['commentCount']
    to_return['n_plays'] = d['stats']['playCount']
    to_return['video_link'] = 'https://www.tiktok.com/@{}/video/{}?lang=en'.format(to_return['user_name'], to_return['video_id'])
    return to_return

dataframe = []

for data in datas:
    for d in data['itemList']:
        dataframe.append(simple_dict(d))

df = pd.DataFrame(dataframe)

df = df.sort_values('video_time', ascending=False).drop_duplicates(subset='video_id').reset_index(drop=True)

df.head()

len(df)

df.to_csv('data/processed/jongraz_videos.csv',index=False)
