==> _workspace/02-process-user-stats.ipynb <==
# Process @jongraz tiktok stats for Noodle videos

import pandas as pd
import numpy as np
import altair as alt

pd.options.display.max_columns = 100
pd.options.display.max_rows = 1000
pd.options.display.max_colwidth = None

### Read in our hand collected list of Noodle videos

src = pd.read_csv(
    "https://docs.google.com/spreadsheets/d/e/2PACX-1vQz8eEafXnIeKsNvGL8M56Ia6vu89JDCjfJ22ORl65So4kRYtmTbwmqLjE223fe4mjcZqkrs-KT6j8z/pub?\
gid=0&single=true&output=csv"
)

### Create an ID field from the URL

src['id'] = src['url'].str.replace('https://www.tiktok.com/@jongraz/video/', "", regex=False).astype(str)

### Make a list for subsetting the dataframe later

bones_videos = src['id'].to_list()

### Have to run this manually -- for now

# Get cookie and run this...

# !python '/Users/stiles/github/noodle-tracker/_workspace/grab_user_stats.py'

---

### Read scraped stats data

df = pd.read_csv('data/processed/jongraz_videos.csv', dtype={'video_id':str})

### Clean update dates, times

df['timestamp'] = pd.to_datetime(df['video_time'], unit='s')
df['date'] = pd.to_datetime(df['video_time'], unit='s').dt.date
df['month'] = pd.to_datetime(df['video_time'], unit='s').dt.month
df['weekday'] = pd.to_datetime(df['video_time'], unit='s').dt.day_name()
df['time'] = pd.to_datetime(df['video_time'], unit='s').dt.time
df['weekstart_timestamp'] = pd.to_datetime(df['video_time'], unit='s') - pd.to_datetime(df['video_time'], unit='s').dt.weekday * np.timedelta64(1, 'D')
df['weekstart'] = df['weekstart_timestamp'].dt.date
df['updated_date'] = pd.to_datetime("today")

### How many videos? 

len(df)

### Select only "bones" videos from @jongraz's feed

bones_df = df[df['video_id'].isin(bones_videos)]

### How many are Noodle videos? 

len(bones_df)

bones_df.head()

---

### Download cover art

# for v,u in zip(bones_df['video_cover'], bones_df['video_id']):
#     !wget {v} -O ../assets/images/noodle_covers/{u}

---

### Weekly plays

weekly_plays = bones_df[bones_df['timestamp'] > '2021/07/31'].groupby(['weekstart']).agg({'n_plays':sum}).reset_index()
weekly_plays.dtypes

weekly_plays['weekstart'] = pd.to_datetime(weekly_plays['weekstart'])

alt.Chart(weekly_plays).mark_bar(width=14).encode(
    x='weekstart:T',
    y='n_plays'
).properties(width=630)

---

### Weekly likes

bones_df.head()

weekly_likes = bones_df[bones_df['timestamp'] > '2021/07/31'].groupby(['weekstart']).agg({'n_likes':sum}).reset_index()
weekly_likes.dtypes

weekly_likes['weekstart'] = pd.to_datetime(weekly_likes['weekstart'])

alt.Chart(weekly_likes).mark_bar(width=20).encode(
    x='weekstart:T',
    y='n_likes'
).properties(width=630)

---

### Weekly comments

weekly_comments = bones_df[bones_df['timestamp'] > '2021/07/31'].groupby(['weekstart']).agg({'n_comments':sum}).reset_index()
weekly_comments.dtypes

weekly_comments['weekstart'] = pd.to_datetime(weekly_comments['weekstart'])

alt.Chart(weekly_comments).mark_bar(width=20).encode(
    x='weekstart:T',
    y='n_comments'
).properties(width=630)

---

### Exports

weekly_plays.to_csv("../_data/weekly_plays.csv", index=False)
weekly_plays.to_csv("../assets/data/weekly_plays.csv", index=False)

weekly_likes.to_csv("../_data/weekly_likes.csv", index=False)
weekly_likes.to_csv("../assets/data/weekly_likes.csv", index=False)

weekly_comments.to_csv("../_data/weekly_comments.csv", index=False)
weekly_comments.to_csv("../assets/data/weekly_comments.csv", index=False)
